as: Extract sinks to dedicated package
