+++ pkg/applicationserver/io/web/sink/health.go
@@ -1,4 +1,4 @@
-// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2024 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package web
+package sink
 
 import (
 	"context"
 	"net/http"
 	"time"
 
-	"go.thethings.network/lorawan-stack/v3/pkg/applicationserver/io/web/internal"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -29,74 +28,6 @@ import (
 type HealthStatusRegistry interface {
 	Get(context.Context) (*ttnpb.ApplicationWebhookHealth, error)
 	Set(context.Context, func(*ttnpb.ApplicationWebhookHealth) (*ttnpb.ApplicationWebhookHealth, error)) error
-}
-
-type healthStatusRegistry struct {
-	registry WebhookRegistry
-}
-
-// Get implements HealthStatusRegistry.
-func (reg *healthStatusRegistry) Get(ctx context.Context) (*ttnpb.ApplicationWebhookHealth, error) {
-	ids := internal.WebhookIDFromContext(ctx)
-	web, err := reg.registry.Get(ctx, ids, []string{"health_status"})
-	if err != nil {
-		return nil, err
-	}
-	return web.HealthStatus, nil
-}
-
-// Set implements HealthStatusRegistry.
-func (reg *healthStatusRegistry) Set(
-	ctx context.Context, f func(*ttnpb.ApplicationWebhookHealth) (*ttnpb.ApplicationWebhookHealth, error),
-) error {
-	ids := internal.WebhookIDFromContext(ctx)
-	_, err := reg.registry.Set(
-		ctx,
-		ids,
-		[]string{"health_status"},
-		func(wh *ttnpb.ApplicationWebhook) (*ttnpb.ApplicationWebhook, []string, error) {
-			if wh == nil {
-				// The webhook has been deleted during execution.
-				return nil, nil, nil
-			}
-			updated, err := f(wh.HealthStatus)
-			if err != nil {
-				return nil, nil, err
-			}
-			wh.HealthStatus = updated
-			return wh, []string{"health_status"}, nil
-		},
-	)
-	return err
-}
-
-// NewHealthStatusRegistry constructs a HealthStatusRegistry on top of the provided WebhookRegistry.
-func NewHealthStatusRegistry(registry WebhookRegistry) HealthStatusRegistry {
-	return &healthStatusRegistry{registry}
-}
-
-type cachedHealthStatusRegistry struct {
-	registry HealthStatusRegistry
-}
-
-// Get implements HealthStatusRegistry.
-func (reg *cachedHealthStatusRegistry) Get(ctx context.Context) (*ttnpb.ApplicationWebhookHealth, error) {
-	if h, ok := internal.WebhookHealthFromContext(ctx); ok {
-		return h, nil
-	}
-	return reg.registry.Get(ctx)
-}
-
-// Set implements HealthStatusRegistry.
-func (reg *cachedHealthStatusRegistry) Set(
-	ctx context.Context, f func(*ttnpb.ApplicationWebhookHealth) (*ttnpb.ApplicationWebhookHealth, error),
-) error {
-	return reg.registry.Set(ctx, f)
-}
-
-// NewCachedHealthStatusRegistry constructs a HealthStatusRegistry which allows the Get response to be cached.
-func NewCachedHealthStatusRegistry(registry HealthStatusRegistry) HealthStatusRegistry {
-	return &cachedHealthStatusRegistry{registry}
 }
 
 type healthCheckSink struct {
