+++ pkg/webui/locales/ja.json
@@ -1806,16 +1806,16 @@
   "error:pkg/applicationserver/io/web/redis:invalid_fieldmask": "フィールドマスク無効",
   "error:pkg/applicationserver/io/web/redis:invalid_identifiers": "識別子無効",
   "error:pkg/applicationserver/io/web/redis:read_only_field": "読み込み専用フィールド `{field}`",
+  "error:pkg/applicationserver/io/web/sink:request": "リクエスト失敗（ステータス `{code}`）",
+  "error:pkg/applicationserver/io/web/sink:webhook_disabled": "Webhookが無効化されています",
   "error:pkg/applicationserver/io/web:decode_body": "デコードされた本体データ",
   "error:pkg/applicationserver/io/web:fetch": "読み取り失敗",
   "error:pkg/applicationserver/io/web:format_not_found": "フォーマット `{format}` が見つかりません",
   "error:pkg/applicationserver/io/web:parse_file": "ファイルをパースできません",
   "error:pkg/applicationserver/io/web:rate_limit_exceeded": "レート制限を超えました",
   "error:pkg/applicationserver/io/web:read_body": "読み取られた本体データ",
-  "error:pkg/applicationserver/io/web:request": "リクエスト失敗（ステータス `{code}`）",
   "error:pkg/applicationserver/io/web:template_not_found": "テンプレート `{template_id}` が見つかりません",
   "error:pkg/applicationserver/io/web:validate_body": "本文を検証する",
-  "error:pkg/applicationserver/io/web:webhook_disabled": "Webhookが無効化されています",
   "error:pkg/applicationserver/io/web:webhook_not_found": "Webhookが見つかりません",
   "error:pkg/applicationserver/io:buffer_full": "バッファがいっぱいです",
   "error:pkg/applicationserver/metadata/redis:cache_miss": "キャッシュミス",
